refactor(cli): extract runtime execute handler
